+++ src/data/types/resourceTypes.ts
@@ -1,0 +1,4 @@
+import type { Sentences, Words, Texts } from "@/data/interfaces/ResourcesData";
+
+export type ResourceType = "words" | "sentences" | "texts";
+export type ResourceData = Words | Sentences | Texts;
